pyproject.toml
[tool.poetry]
name = "finetuning-encoders"
version = "0.1.0"
description = ""
authors = ["Arda Can Aras <ardaaras99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
transformers = "4.18.0"
jupyter = "^1.0.0"
numpy = "^1.19.2"
scikit-learn = "^1.2.0"
pandas = "^1.5.2"
torch = "2.0.0"
datasets = "^2.17.1"
baseline-text-graphs = { git = "https://github.com/ardaaras99/baseline-text-graphs.git", rev = "main" }
click = "^8.1.7"
python-dotenv = "^1.0.1"
wandb = { extras = ["sweeps"], version = "^0.15.12" }


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"
ruff = "^0.1.9"

[tool.ruff]
line-length = 300

[tool.ruff.format]
# Like Black
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
extend-select = [
    "A",  # flake8-builtins
    "B",  # flake8-bugbear
    "S",  # flake8-bandit
    "I",  # isort
    "UP", # pyupgrade
]

[tool.ruff.lint.extend-per-file-ignores]
"*" = ["S101", "S301", "S311"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
